+++ Sources/Sentry/include/SentryProfilingConditionals.h
@@ -1,0 +1,1 @@
+HybridPublic/SentryProfilingConditionals.h