+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Breaking Changes
+
+- Removes SentryEnvelopeItemHeader from the public API (#6870)
+
+### Improvements
+
+- Change macOS deployment target to 10.14 (#6873)
+
 ## 9.0.0-rc.0
 
 ### Breaking Changes